remove roles with implementation, add as a dependency
The roles Pod::Simple::Role::WithHighlightConfig and
Pod::Simple::Role::XHTML::WithHighlightConfig have moved into the
MetaCPAN-Pod-HTML dist. This dist only now exists for backwards
compatibility.

The main module of this dist was never really meant to be used, but just
seemed a reasonable thing to provide and a name that could be linked to.
The primary thing meant to be used was the role
Pod::Simple::Role::XHTML::WithHighlightConfig. That role has now been moved
into the MetaCPAN-Pod-HTML dist, so that all of MetaCPAN's Pod handling
extensions can be maintained together.

This dist now contains essentially zero code, so hopefully will never
need any further updates or releases.